requirements.txt
python-dateutil
emoji>=1.7.0